+++ examples/table/demos/merge-cells.vue
@@ -25,7 +25,6 @@ const columns = [
   {
     align: 'left',
     width: '100',
-    minWidth: '100',
     className: 'test',
     colKey: 'platform',
     title: '平台',
@@ -33,7 +32,6 @@ const columns = [
   {
     align: 'left',
     width: '100',
-    minWidth: '100',
     className: 'row',
     colKey: 'type',
     title: '类型',
@@ -41,7 +39,6 @@ const columns = [
   {
     align: 'left',
     width: '100',
-    minWidth: '100',
     className: 'test4',
     colKey: 'default',
     title: '默认值',
@@ -49,7 +46,6 @@ const columns = [
   {
     align: 'left',
     width: '100',
-    minWidth: '100',
     className: 'test3',
     colKey: 'needed',
     title: '是否必传',
@@ -57,7 +53,6 @@ const columns = [
   {
     align: 'left',
     width: '100',
-    minWidth: '100',
     className: 'row',
     colKey: 'description',
     title: '说明',