+++ src/date-picker/hooks/useTableData.tsx
@@ -1,6 +1,6 @@
 import { useConfig } from '../../hooks/useConfig';
 
-import { getWeeks, getYears, getMonths, flagActive } from '../../_common/js/date-picker/utils';
+import { getWeeks, getYears, getMonths, getQuarters, flagActive } from '../../_common/js/date-picker/utils';
 
 export default function useTableData(props: any) {
   // 国际化文本初始化
@@ -12,11 +12,17 @@ export default function useTableData(props: any) {
     disableDate: props.disableDate,
     firstDayOfWeek: props.firstDayOfWeek ?? global.value.firstDayOfWeek,
     monthLocal: global.value.months,
+    quarterLocal: global.value.quarters,
+    showWeekOfYear: props.mode === 'week',
   };
 
   let data: Array<any> = [];
   if (props.mode === 'date') {
     data = getWeeks({ year: props.year, month: props.month }, options);
+  } else if (props.mode === 'week') {
+    data = getWeeks({ year: props.year, month: props.month }, options);
+  } else if (props.mode === 'quarter') {
+    data = getQuarters(props.year, options);
   } else if (props.mode === 'month') {
     data = getMonths(props.year, options);
   } else if (props.mode === 'year') {