+++ src/date-picker/panel/SinglePanel.tsx
@@ -1,10 +1,9 @@
 import { defineComponent, PropType, computed } from 'vue';
-import dayjs from 'dayjs';
 import { useConfig, usePrefixClass } from '../../hooks/useConfig';
 import TPanelContent from './PanelContent';
 import TExtraContent from './ExtraContent';
-import { TdDatePickerProps, DateValue } from '../type';
-import { getDefaultFormat } from '../hooks/useFormat';
+import { TdDatePickerProps } from '../type';
+import { getDefaultFormat, parseToDayjs } from '../hooks/useFormat';
 import useTableData from '../hooks/useTableData';
 import useDisableDate from '../hooks/useDisableDate';
 
@@ -51,10 +50,6 @@ export default defineComponent({
       enableTimePicker: props.enableTimePicker,
     });
 
-    // 兼容数据格式不标准场景 YYYY-MM-D
-    const formatDate = (newDate: DateValue, format: string) =>
-      dayjs(newDate).isValid() ? dayjs(newDate).toDate() : dayjs(newDate, format).toDate();
-
     const disableDateOptions = computed(() =>
       useDisableDate({
         format,
@@ -68,7 +63,7 @@ export default defineComponent({
         year: props.year,
         month: props.month,
         mode: props.mode,
-        start: props.value ? formatDate(props.value, format) : undefined,
+        start: props.value ? parseToDayjs(props.value, format).toDate() : undefined,
         firstDayOfWeek: props.firstDayOfWeek || global.value.firstDayOfWeek,
         ...disableDateOptions.value,
       }),
@@ -76,6 +71,7 @@ export default defineComponent({
 
     const panelContentProps = computed(() => ({
       format,
+      value: props.value,
       mode: props.mode,
       year: props.year,
       month: props.month,