+++ src/date-picker/props.ts
@@ -12,12 +12,12 @@ export default {
   allowInput: Boolean,
   /** 是否显示清除按钮 */
   clearable: Boolean,
-  /** 是否禁用组件 */
-  disabled: Boolean,
   /** 禁用日期，示例：['A', 'B'] 表示日期 A 和日期 B 会被禁用。`{ from: 'A', to: 'B' }` 表示在 A 到 B 之间的日期会被禁用。`{ before: 'A', after: 'B' }` 表示在 A 之前和在 B 之后的日期都会被禁用。其中 A = '2021-01-01'，B = '2021-02-01'。值类型为 Function 则表示返回值为 true 的日期会被禁用 */
   disableDate: {
     type: [Object, Array, Function] as PropType<TdDatePickerProps['disableDate']>,
   },
+  /** 是否禁用组件 */
+  disabled: Boolean,
   /** 是否显示时间选择 */
   enableTimePicker: Boolean,
   /** 第一天从星期几开始 */
@@ -31,7 +31,7 @@ export default {
   /** 用于格式化日期，全局配置默认为：'YYYY-MM-DD'，[详细文档](https://day.js.org/docs/en/display/format) */
   format: {
     type: String,
-    default: '',
+    default: undefined,
   },
   /** 透传给输入框（Input）组件的参数 */
   inputProps: {
@@ -43,7 +43,7 @@ export default {
     default: 'date' as TdDatePickerProps['mode'],
     validator(val: TdDatePickerProps['mode']): boolean {
       if (!val) return true;
-      return ['year', 'month', 'date'].includes(val);
+      return ['year', 'quarter', 'month', 'week', 'date'].includes(val);
     },
   },
   /** 占位符 */
@@ -72,6 +72,14 @@ export default {
       return ['left', 'top', 'right', 'bottom'].includes(val);
     },
   },
+  /** 输入框状态 */
+  status: {
+    type: String as PropType<TdDatePickerProps['status']>,
+    validator(val: TdDatePickerProps['status']): boolean {
+      if (!val) return true;
+      return ['default', 'success', 'warning', 'error'].includes(val);
+    },
+  },
   /** 用于自定义组件后置图标 */
   suffixIcon: {
     type: Function as PropType<TdDatePickerProps['suffixIcon']>,
@@ -79,6 +87,10 @@ export default {
   /** 透传 TimePicker 组件属性 */
   timePickerProps: {
     type: Object as PropType<TdDatePickerProps['timePickerProps']>,
+  },
+  /** 输入框下方提示文本，会根据不同的 `status` 呈现不同的样式 */
+  tips: {
+    type: [String, Function] as PropType<TdDatePickerProps['tips']>,
   },
   /** 选中值 */
   value: {