+++ src/pagination/pagination.tsx
@@ -13,6 +13,7 @@ import { TdPaginationProps } from '../pagination/type';
 import { useConfig, usePrefixClass } from '../hooks/useConfig';
 import TInputNumber from '../input-number';
 import { Option, Select } from '../select';
+import TInputAdornment from '../input-adornment';
 import props from './props';
 import usePaginationClasses from './usePaginationClasses';
 import useMoreAction from './useMoreAction';
@@ -198,8 +199,28 @@ export default defineComponent({
     };
 
     return () => {
-      const { total, pageSizeOptions, size, disabled, showJumper, showPageSize } = props;
+      const { total, pageSizeOptions, size, disabled, showPageSize } = props;
       if (pageCount.value < 1) return null;
+
+      const Jumper = (
+        <div class={CLASS_MAP.jumperClass.value}>
+          {t(global.value.jumpTo)}
+          <TInputAdornment append={`/ ${pageCount.value} ${t(global.value.page)}`}>
+            <TInputNumber
+              class={CLASS_MAP.jumperInputClass.value}
+              v-model={jumpIndex.value}
+              onBlur={onJumperChange}
+              onEnter={onJumperChange}
+              max={pageCount.value}
+              min={min}
+              size={size}
+              disabled={disabled}
+              theme="normal"
+              placeholder=""
+            />
+          </TInputAdornment>
+        </div>
+      );
 
       return (
         <div class={CLASS_MAP.paginationClass.value}>
@@ -285,17 +306,7 @@ export default defineComponent({
             </ul>
           ) : null}
           {/* 极简版 */}
-          {props.showPageNumber && props.theme === 'simple' ? (
-            <Select
-              size={size}
-              value={innerCurrent}
-              disabled={disabled}
-              class={CLASS_MAP.simpleClass.value}
-              autoWidth={true}
-              onChange={(value) => toPage(value as number)}
-              options={pageCountOption.value}
-            />
-          ) : null}
+          {props.theme === 'simple' && Jumper}
           {/* 向后按钮 */}
           {props.showPreviousAndNextBtn ? (
             <div
@@ -316,25 +327,8 @@ export default defineComponent({
               <PageLastIcon />
             </div>
           ) : null}
-          {/* 跳转 */}
-          {showJumper ? (
-            <div class={CLASS_MAP.jumperClass.value}>
-              {t(global.value.jumpTo)}
-              <TInputNumber
-                class={CLASS_MAP.jumperInputClass.value}
-                v-model={jumpIndex.value}
-                onBlur={onJumperChange}
-                onEnter={onJumperChange}
-                max={pageCount.value}
-                min={min}
-                size={size}
-                disabled={disabled}
-                theme="normal"
-                placeholder=""
-              />
-              {t(global.value.page)}
-            </div>
-          ) : null}
+          {/* 快速跳转 */}
+          {props.theme === 'default' && props.showJumper && Jumper}
         </div>
       );
     };